Move jaialogs python files into server directory
